LIBILS-386. Modified database_credentials to return setting dicts
Modified database_credentials to return dicts for the application
and test database settings.

This more clearly separates the application and test database settings,
and allows an error message to be printed (and the system exited) if
the application database settings have not been added when the
application is run (but tests can still be run, if a test configuration
is provided).

https://issues.umd.edu/browse/LIBILS-386