LIBILS-390. Added "database_session" context manager
Following the guidance in

https://docs.sqlalchemy.org/en/13/orm/session_basics.html

created "database_session" and "test_database_session" context managers
that enables a database connection to be used with the Python "with"
idiom.

The "test_database_session" runs in a transaction, so that any changes
(even if they are committed) can be rolled back.

The "database_session" will actually commit data to the database,
unless an exception is raised, in which case the changes will be
rolled back.

https://issues.umd.edu/browse/LIBILS-390