LIBILS-370. Updated tests to use negative processing cycle ids
In a local development environment, the database may be shared between
the application and the tests.

In order to ensure that having actual data in the database will not
interfere with running the tests, the tests should use negative
processing cycle ids. Since negative processing cycle ids will never
be used for application data, this ensures that the application data
and test data will not be in conflict.

When a test actually uses the database, the test_database_session is
designed to roll back the changes after the test is completed. So the
negative processing cycle ids should never be seen in the database
once the test completes.

https://issues.umd.edu/browse/LIBILS-370