LIBILS-370. job_info creation should occur only once in Stage 1
Fixed an issue where a new job_info object was being created for
every file being loaded.

The intended use of job_info is to provide a single "processing_cycle_id"
for a single run of the complete ETL script, so creating a new one
for each file being loaded was incorrect.

https://issues.umd.edu/browse/LIBILS-370